requirements.txt
# Core dependencies
pdfplumber>=0.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0

# Optional pero recomendado
tabula-py>=2.8.0  # Fallback para parsing de tablas
camelot-py[cv]>=0.11.0  # Alternativa para PDFs complejos

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
